add jenkins file tests

+++ backend/tests/setup.js
@@ -24,8 +24,11 @@ beforeAll(async () => {
 afterEach(async () => {
   const collections = mongoose.connection.collections;
   for (const key in collections) {
-    const collection = collections[key];
-    await collection.deleteMany();
+    if (Object.prototype.hasOwnProperty.call(collections, key)) {
+      // eslint-disable-next-line security/detect-object-injection
+      const collection = collections[key];
+      await collection.deleteMany();
+    }
   }
 });
